Add forecast weather displaying

+++ androidApp/src/main/java/com/simple/weather/app/android/data/model/ForecastDay.kt
@@ -8,7 +8,7 @@ data class ForecastDay(
     @SerialName("date")
     val date: String, // 2021-09-14
     @SerialName("date_epoch")
-    val dateEpoch: Int, // 1631577600
+    val dateEpoch: Long, // 1631577600
     @SerialName("day")
     val day: ForecastDayData,
     @SerialName("astro")
